+++ forms-flow-api/src/formsflow_api/resources/roles.py
@@ -5,10 +5,11 @@ from http import HTTPStatus
 from flask import request
 from flask_restx import Namespace, Resource, fields
 from formsflow_api_utils.utils import (
-    ADMIN,
     CREATE_DESIGNS,
     CREATE_FILTERS,
     MANAGE_ALL_FILTERS,
+    MANAGE_ROLES,
+    MANAGE_USERS,
     PERMISSION_DETAILS,
     VIEW_DESIGNS,
     VIEW_FILTERS,
@@ -55,12 +56,13 @@ class KeycloakRolesResource(Resource):
     @staticmethod
     @auth.has_one_of_roles(
         [
-            ADMIN,
             CREATE_DESIGNS,
             MANAGE_ALL_FILTERS,
             CREATE_FILTERS,
             VIEW_FILTERS,
             VIEW_DESIGNS,
+            MANAGE_USERS,
+            MANAGE_ROLES,
         ]
     )
     @profiletime
@@ -92,7 +94,7 @@ class KeycloakRolesResource(Resource):
         return response, HTTPStatus.OK
 
     @staticmethod
-    @auth.has_one_of_roles([ADMIN])
+    @auth.has_one_of_roles([MANAGE_ROLES])
     @profiletime
     @API.doc(
         responses={
@@ -120,7 +122,7 @@ class KeycloakRolesResourceById(Resource):
     """Resource to manage keycloak roles/groups by id."""
 
     @staticmethod
-    @auth.has_one_of_roles([ADMIN])
+    @auth.has_one_of_roles([MANAGE_ROLES])
     @profiletime
     @API.doc(
         responses={
@@ -137,7 +139,7 @@ class KeycloakRolesResourceById(Resource):
         return response, HTTPStatus.OK
 
     @staticmethod
-    @auth.has_one_of_roles([ADMIN])
+    @auth.has_one_of_roles([MANAGE_ROLES])
     @profiletime
     @API.doc(
         responses={
@@ -152,7 +154,7 @@ class KeycloakRolesResourceById(Resource):
         return {"message": "Deleted successfully."}, HTTPStatus.OK
 
     @staticmethod
-    @auth.has_one_of_roles([ADMIN])
+    @auth.has_one_of_roles([MANAGE_ROLES])
     @profiletime
     @API.doc(
         responses={
@@ -175,7 +177,7 @@ class Permissions(Resource):
     """Resource to list."""
 
     @staticmethod
-    @auth.has_one_of_roles([ADMIN])
+    @auth.has_one_of_roles([MANAGE_ROLES])
     @profiletime
     @API.doc(
         responses={